requirements.txt
python-vlc
asyncio
PyQt5
bilibili-api-python
requests